🛠 Fix statement

+++ cmds/purge.js
@@ -9,25 +9,28 @@ module.exports = {
     ),
   usage: "amount: <number>",
   async execute(interaction) {
-    const amount = interaction.options.getInteger("amount");
+    if (interaction) {
+      const amount = interaction.options.getInteger("amount");
 
-    if (amount <= 1 || amount > 100) {
+      if (amount <= 1 || amount > 100) {
+        return interaction.reply({
+          content: "You need to input a number between 1 and 99.",
+          ephemeral: true,
+        });
+      }
+      await interaction.channel.bulkDelete(amount, true).catch((error) => {
+        console.error(error);
+        interaction.reply({
+          content:
+            "There was an error trying to purge messages in this channel!",
+          ephemeral: true,
+        });
+      });
+
       return interaction.reply({
-        content: "You need to input a number between 1 and 99.",
+        content: `Successfully purged \`${amount}\` messages.`,
         ephemeral: true,
       });
     }
-    await interaction.channel.bulkDelete(amount, true).catch((error) => {
-      console.error(error);
-      interaction.reply({
-        content: "There was an error trying to purge messages in this channel!",
-        ephemeral: true,
-      });
-    });
-
-    return interaction.reply({
-      content: `Successfully purged \`${amount}\` messages.`,
-      ephemeral: true,
-    });
   },
 };
